requirements.txt
maafw >=2.4.0
Pillow
